HOL-Light: Add functional correctness proof for rej_uniform
This commit imports John Harrison's HOL-Light specification & proof
of functional correctness for the AArch64 Neon implementation of
the rejection sampling used in mlkem-native.

Since not all instructions needed in the code are part of the s2n-bignum
repository yet, John's development branch of s2n-bignum is used for the
time being.

Also, various instructions have been changed to use the X-form rather
than the W-form of general purpose registers. This is to work around
unmodelled AArch64 instructions in s2n-bignum.

For all the complexity in the AArch64 assembly, the functional
specification for rej_uniform is remarkably simple: It's effectively
just a filter by `(<) 3329`.

+++ dev/aarch64_opt/src/arith_native_aarch64.h
@@ -70,7 +70,7 @@ void mlk_polyvec_basemul_acc_montgomery_cached_asm_k4(int16_t *r,
                                                       const int16_t *b_cache);
 
 #define mlk_rej_uniform_asm MLK_NAMESPACE(rej_uniform_asm)
-unsigned mlk_rej_uniform_asm(int16_t *r, const uint8_t *buf, unsigned buflen,
+uint64_t mlk_rej_uniform_asm(int16_t *r, const uint8_t *buf, unsigned buflen,
                              const uint8_t *table);
 
 #endif /* !MLK_DEV_AARCH64_OPT_SRC_ARITH_NATIVE_AARCH64_H */
